+++ packages/terra-form-select/src/MultiSelectField.jsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { injectIntl, intlShape } from 'react-intl';
 import Field from 'terra-form-field';
-import DropdownSelect from './DropdownSelect';
+import MultiSelect from './MultiSelect';
 import OptGroup from './shared/_OptGroup';
 import Option from './shared/_Option';
 
 const propTypes = {
-  /**
-   * Whether a clear option is available to clear the selection.
-   */
-  allowClear: PropTypes.bool,
   /**
    * The select options.
    */
@@ -21,7 +18,7 @@ const propTypes = {
   /**
    * The default value of the select.
    */
-  defaultValue: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  defaultValue: PropTypes.arrayOf(PropTypes.string),
   /**
    * Whether the input is disabled.
    */
@@ -39,6 +36,11 @@ const propTypes = {
    */
   hideRequired: PropTypes.bool,
   /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: intlShape.isRequired,
+  /**
    * Whether the field is displayed inline. Displays block by default.
    */
   isInline: PropTypes.bool,
@@ -55,6 +57,10 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   labelAttrs: PropTypes.object,
+  /**
+   * The maximum number of options that can be selected. A value less than 2 will be ignored.
+   */
+  maxSelectionCount: PropTypes.number,
   /**
    * Set the max-width of a field using `length` or `%`.  Best practice recommendation to never exceed
    * a rendered value of 1020px. _(Note: Providing custom inline styles will take precedence.)_
@@ -88,11 +94,10 @@ const propTypes = {
   /**
    * The value of the select.
    */
-  value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  value: PropTypes.arrayOf(PropTypes.string),
 };
 
 const defaultProps = {
-  allowClear: false,
   children: undefined,
   defaultValue: undefined,
   disabled: false,
@@ -103,6 +108,7 @@ const defaultProps = {
   isInvalid: false,
   isLabelHidden: false,
   labelAttrs: {},
+  maxSelectionCount: undefined,
   maxWidth: undefined,
   onChange: undefined,
   placeholder: undefined,
@@ -112,19 +118,20 @@ const defaultProps = {
   value: undefined,
 };
 
-const DropdownSelectField = ({
-  allowClear,
+const MultiSelectField = ({
   children,
   defaultValue,
   disabled,
   error,
   help,
   hideRequired,
+  intl,
   isInline,
   isInvalid,
   isLabelHidden,
   label,
   labelAttrs,
+  maxSelectionCount,
   maxWidth,
   onChange,
   placeholder,
@@ -135,6 +142,23 @@ const DropdownSelectField = ({
   value,
   ...customProps
 }) => {
+  let helpText = help;
+  if (maxSelectionCount !== undefined && maxSelectionCount >= 2) {
+    const limitSelectionText = intl.formatMessage({ id: 'Terra.form.select.maxSelectionHelp' }, { text: maxSelectionCount });
+
+    if (help) {
+      helpText = (
+        <span>
+          { limitSelectionText }
+          { ' ' }
+          { help }
+        </span>
+      );
+    } else {
+      helpText = limitSelectionText;
+    }
+  }
+
   let ariaDescriptionIds;
 
   if (help && error && isInvalid) {
@@ -155,7 +179,7 @@ const DropdownSelectField = ({
       label={label}
       labelAttrs={labelAttrs}
       error={error}
-      help={help}
+      help={helpText}
       hideRequired={hideRequired}
       required={required}
       showOptional={showOptional}
@@ -165,30 +189,30 @@ const DropdownSelectField = ({
       htmlFor={selectId}
       maxWidth={maxWidth}
     >
-      <DropdownSelect
+      <MultiSelect
         {...selectAttrs}
         ariaLabel={label}
-        allowClear={allowClear}
         aria-describedby={ariaDescriptionIds}
         disabled={selectAttrs.disabled || disabled}
         id={selectId}
         isInvalid={isInvalid}
         defaultValue={defaultValue}
+        maxSelectionCount={maxSelectionCount !== undefined && maxSelectionCount < 2 ? undefined : maxSelectionCount}
         onChange={onChange}
         placeholder={placeholder}
         required={required}
         value={value}
       >
         {children}
-      </DropdownSelect>
+      </MultiSelect>
     </Field>
   );
 };
 
-DropdownSelectField.propTypes = propTypes;
-DropdownSelectField.defaultProps = defaultProps;
+MultiSelectField.propTypes = propTypes;
+MultiSelectField.defaultProps = defaultProps;
 
-DropdownSelectField.Option = Option;
-DropdownSelectField.OptGroup = OptGroup;
+MultiSelectField.Option = Option;
+MultiSelectField.OptGroup = OptGroup;
 
-export default DropdownSelectField;
+export default injectIntl(MultiSelectField);