+++ packages/terra-form-select/src/SingleSelectField.jsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { injectIntl, intlShape } from 'react-intl';
 import Field from 'terra-form-field';
-import MultipleSelect from './MultipleSelect';
+import SingleSelect from './SingleSelect';
 import OptGroup from './shared/_OptGroup';
 import Option from './shared/_Option';
 
 const propTypes = {
+  /**
+   * Whether a clear option is available to clear the selection.
+   */
+  allowClear: PropTypes.bool,
   /**
    * The select options.
    */
@@ -18,7 +21,7 @@ const propTypes = {
   /**
    * The default value of the select.
    */
-  defaultValue: PropTypes.arrayOf(PropTypes.string),
+  defaultValue: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   /**
    * Whether the input is disabled.
    */
@@ -36,11 +39,6 @@ const propTypes = {
    */
   hideRequired: PropTypes.bool,
   /**
-   * @private
-   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
-   */
-  intl: intlShape.isRequired,
-  /**
    * Whether the field is displayed inline. Displays block by default.
    */
   isInline: PropTypes.bool,
@@ -57,10 +55,6 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   labelAttrs: PropTypes.object,
-  /**
-   * The maximum number of options that can be selected. A value less than 2 will be ignored.
-   */
-  maxSelectionCount: PropTypes.number,
   /**
    * Set the max-width of a field using `length` or `%`.  Best practice recommendation to never exceed
    * a rendered value of 1020px. _(Note: Providing custom inline styles will take precedence.)_
@@ -94,10 +88,11 @@ const propTypes = {
   /**
    * The value of the select.
    */
-  value: PropTypes.arrayOf(PropTypes.string),
+  value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
 };
 
 const defaultProps = {
+  allowClear: false,
   children: undefined,
   defaultValue: undefined,
   disabled: false,
@@ -108,7 +103,6 @@ const defaultProps = {
   isInvalid: false,
   isLabelHidden: false,
   labelAttrs: {},
-  maxSelectionCount: undefined,
   maxWidth: undefined,
   onChange: undefined,
   placeholder: undefined,
@@ -118,20 +112,19 @@ const defaultProps = {
   value: undefined,
 };
 
-const MultipleSelectField = ({
+const SingleSelectField = ({
+  allowClear,
   children,
   defaultValue,
   disabled,
   error,
   help,
   hideRequired,
-  intl,
   isInline,
   isInvalid,
   isLabelHidden,
   label,
   labelAttrs,
-  maxSelectionCount,
   maxWidth,
   onChange,
   placeholder,
@@ -142,23 +135,6 @@ const MultipleSelectField = ({
   value,
   ...customProps
 }) => {
-  let helpText = help;
-  if (maxSelectionCount !== undefined && maxSelectionCount >= 2) {
-    const limitSelectionText = intl.formatMessage({ id: 'Terra.form.select.maxSelectionHelp' }, { text: maxSelectionCount });
-
-    if (help) {
-      helpText = (
-        <span>
-          { limitSelectionText }
-          { ' ' }
-          { help }
-        </span>
-      );
-    } else {
-      helpText = limitSelectionText;
-    }
-  }
-
   let ariaDescriptionIds;
 
   if (help && error && isInvalid) {
@@ -179,7 +155,7 @@ const MultipleSelectField = ({
       label={label}
       labelAttrs={labelAttrs}
       error={error}
-      help={helpText}
+      help={help}
       hideRequired={hideRequired}
       required={required}
       showOptional={showOptional}
@@ -189,30 +165,30 @@ const MultipleSelectField = ({
       htmlFor={selectId}
       maxWidth={maxWidth}
     >
-      <MultipleSelect
+      <SingleSelect
         {...selectAttrs}
         ariaLabel={label}
+        allowClear={allowClear}
         aria-describedby={ariaDescriptionIds}
         disabled={selectAttrs.disabled || disabled}
         id={selectId}
         isInvalid={isInvalid}
         defaultValue={defaultValue}
-        maxSelectionCount={maxSelectionCount !== undefined && maxSelectionCount < 2 ? undefined : maxSelectionCount}
         onChange={onChange}
         placeholder={placeholder}
         required={required}
         value={value}
       >
         {children}
-      </MultipleSelect>
+      </SingleSelect>
     </Field>
   );
 };
 
-MultipleSelectField.propTypes = propTypes;
-MultipleSelectField.defaultProps = defaultProps;
+SingleSelectField.propTypes = propTypes;
+SingleSelectField.defaultProps = defaultProps;
 
-MultipleSelectField.Option = Option;
-MultipleSelectField.OptGroup = OptGroup;
+SingleSelectField.Option = Option;
+SingleSelectField.OptGroup = OptGroup;
 
-export default injectIntl(MultipleSelectField);
+export default SingleSelectField;