+++ packages/terra-form-select/src/shared/_SelectUtil.js
@@ -15,7 +15,7 @@ class SelectUtil {
       return null;
     }
 
-    if (SelectUtil.allowsMultipleSelections(props)) {
+    if (SelectUtil.allowsMultiSelections(props)) {
       // Flatten allows converting a string default into an array.
       return defaultValue ? [defaultValue].flatten() : [];
     }
@@ -50,7 +50,7 @@ class SelectUtil {
    * @param {Object} props - The component props.
    * @return {boolean} - True if the variant allows multiple selections.
    */
-  static allowsMultipleSelections(props) {
+  static allowsMultiSelections(props) {
     return props.variant === Variants.MULTIPLE || props.variant === Variants.TAG;
   }
 
@@ -62,7 +62,7 @@ class SelectUtil {
    * @return {array|number|string} - The value resulting from the selection of an option
    */
   static select(props, state, value) {
-    if (SelectUtil.allowsMultipleSelections(props)) {
+    if (SelectUtil.allowsMultiSelections(props)) {
       return [...SelectUtil.value(props, state), value];
     }
     return value;