+++ packages/terra-form-select/src/terra-dev-site/doc/example/multiple/Controlled.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import MultipleSelect from 'terra-form-select/lib/MultipleSelect';
+import MultiSelect from 'terra-form-select/lib/MultiSelect';
 import classNames from 'classnames/bind';
 import styles from '../FormSelectDocCommon.module.scss';
 
@@ -10,13 +10,13 @@ const ControlledExample = () => {
   const [value, setValue] = useState(['blue']);
 
   return (
-    <MultipleSelect placeholder="Select a color" value={value} onChange={setValue} className={cx('form-select')}>
-      <MultipleSelect.Option value="blue" display="Blue" />
-      <MultipleSelect.Option value="green" display="Green" />
-      <MultipleSelect.Option value="purple" display="Purple" />
-      <MultipleSelect.Option value="red" display="Red" />
-      <MultipleSelect.Option value="violet" display="Violet" />
-    </MultipleSelect>
+    <MultiSelect placeholder="Select a color" value={value} onChange={setValue} className={cx('form-select')}>
+      <MultiSelect.Option value="blue" display="Blue" />
+      <MultiSelect.Option value="green" display="Green" />
+      <MultiSelect.Option value="purple" display="Purple" />
+      <MultiSelect.Option value="red" display="Red" />
+      <MultiSelect.Option value="violet" display="Violet" />
+    </MultiSelect>
   );
 };