+++ packages/terra-form-select/src/terra-dev-site/doc/example/multiple/MultiSelect.jsx
@@ -6,15 +6,14 @@ import styles from '../FormSelectDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const DisabledExample = () => (
-  <MultiSelect placeholder="Select a color" disabled className={cx('form-select')} defaultValue={['red', 'white', 'blue']}>
+const MultiSelectExample = () => (
+  <MultiSelect placeholder="Select a color" className={cx('form-select')}>
     <MultiSelect.Option value="blue" display="Blue" />
     <MultiSelect.Option value="green" display="Green" />
     <MultiSelect.Option value="purple" display="Purple" />
     <MultiSelect.Option value="red" display="Red" />
     <MultiSelect.Option value="violet" display="Violet" />
-    <MultiSelect.Option value="white" display="White" />
   </MultiSelect>
 );
 
-export default DisabledExample;
+export default MultiSelectExample;