+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/MultiSelectField.08.doc.jsx
@@ -1,18 +1,18 @@
 /* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../../docs/MultipleSelectField.md';
+import ReadMe from '../../../../docs/MultiSelectField.md';
 import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import MultipleSelectFieldSrc from '!raw-loader!../../../../src/MultipleSelectField';
+import MultiSelectFieldSrc from '!raw-loader!../../../../src/MultiSelectField';
 
 // Example Files
-import SelectFieldExample from '../example/multiple/MultipleSelectField';
-import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/MultipleSelectField';
-import RequiredMultipleSelectFieldExample from '../example/multiple/RequiredMultipleSelectField';
-import RequiredMultipleSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/RequiredMultipleSelectField';
+import SelectFieldExample from '../example/multiple/MultiSelectField';
+import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/MultiSelectField';
+import RequiredMultiSelectFieldExample from '../example/multiple/RequiredMultiSelectField';
+import RequiredMultiSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/multiple/RequiredMultiSelectField';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -21,17 +21,17 @@ const DocPage = () => (
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTables={[{
-      componentName: 'Multiple Select Field',
-      componentSrc: MultipleSelectFieldSrc,
+      componentName: 'MultiSelect Field',
+      componentSrc: MultiSelectFieldSrc,
     }]}
     examples={[{
-      title: 'Multiple Select Field Example',
+      title: 'MultiSelect Field Example',
       example: <SelectFieldExample />,
       source: SelectFieldExampleSrc,
     }, {
-      title: 'Required Multiple Select Field Example',
-      example: <RequiredMultipleSelectFieldExample />,
-      source: RequiredMultipleSelectFieldExampleSrc,
+      title: 'Required MultiSelect Field Example',
+      example: <RequiredMultiSelectFieldExample />,
+      source: RequiredMultiSelectFieldExampleSrc,
     }]}
   />
 );