+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SingleSelectField.04.doc.jsx
@@ -1,18 +1,18 @@
 /* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../../docs/DropdownSelectField.md';
+import ReadMe from '../../../../docs/SingleSelectField.md';
 import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import DropdownSelectFieldSrc from '!raw-loader!../../../../src/DropdownSelectField';
+import SingleSelectFieldSrc from '!raw-loader!../../../../src/SingleSelectField';
 
 // Example Files
-import SelectFieldExample from '../example/dropdown/DropdownSelectField';
-import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/dropdown/DropdownSelectField';
-import RequiredDropdownSelectFieldExample from '../example/dropdown/RequiredDropdownSelectField';
-import RequiredDropdownSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/dropdown/RequiredDropdownSelectField';
+import SelectFieldExample from '../example/single/SingleSelectField';
+import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/single/SingleSelectField';
+import RequiredSingleSelectFieldExample from '../example/single/RequiredSingleSelectField';
+import RequiredSingleSelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/single/RequiredSingleSelectField';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -21,17 +21,17 @@ const DocPage = () => (
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTables={[{
-      componentName: 'Dropdown Select Field',
-      componentSrc: DropdownSelectFieldSrc,
+      componentName: 'Single Select Field',
+      componentSrc: SingleSelectFieldSrc,
     }]}
     examples={[{
-      title: 'Dropdown Select Field Example',
+      title: 'Single Select Field Example',
       example: <SelectFieldExample />,
       source: SelectFieldExampleSrc,
     }, {
-      title: 'Required Dropdown Select Field Example',
-      example: <RequiredDropdownSelectFieldExample />,
-      source: RequiredDropdownSelectFieldExampleSrc,
+      title: 'Required Single Select Field Example',
+      example: <RequiredSingleSelectFieldExample />,
+      source: RequiredSingleSelectFieldExampleSrc,
     }]}
   />
 );